Adding backbones to own module. Adding covernet implementation
